+++ E-Sharing/src/ESharing/Client/Model/AdvertisementModel/AdvertisementModelManager.java
@@ -112,7 +112,7 @@ public class AdvertisementModelManager implements AdvertisementModel{
     }
 
     @Override
-    public boolean addRating(int ad_id, int user_id, int rating)
+    public boolean addRating(int ad_id, int user_id, double rating)
     {
         return clientAdvertisement.addRating(ad_id, user_id, rating);
     }
